+++ src/functions/public/Get-DirectorySecurity.ps1
@@ -1,0 +1,37 @@
+function Get-DirectorySecurity {
+    <#
+    .SYNOPSIS
+    Alternative to Get-Acl designed to be as lightweight and flexible as possible
+        Lightweight: Does not return the Path property like Get-Acl does
+        Flexible how?  Was it long paths?  DFS?  Can't remember what didn't work with Get-Acl
+    TEMP NOTE: Get-DirectorySecurity combined with Get-FileSystemAccessRule replaces Get-FolderACE
+    .DESCRIPTION
+    Returns an object for each access control entry instead of a single object for the ACL
+    Excludes inherited permissions by default but allows them to be included with the -IncludeInherited switch parameter
+    .INPUTS
+    None. Pipeline input is not accepted.
+    .OUTPUTS
+    [PSCustomObject]
+    .NOTES
+    Currently only supports Directories but could easily be copied to support files, or Registry or AD providers
+    #>
+
+    param(
+
+        # Path to the directory whose permissions to get
+        [string]$LiteralPath,
+
+        # Include all sections except Audit because it requires admin rights if run on the local system and we want to avoid that requirement
+        [System.Security.AccessControl.AccessControlSections]$Sections = (
+            [System.Security.AccessControl.AccessControlSections]::Access -bor
+            [System.Security.AccessControl.AccessControlSections]::Owner -bor
+            [System.Security.AccessControl.AccessControlSections]::Group)
+
+    )
+
+    [System.Security.AccessControl.DirectorySecurity]::new(
+        $LiteralPath,
+        $Sections
+    )
+
+}