+++ src/functions/public/Get-FileSystemAccessRule.ps1
@@ -1,0 +1,69 @@
+function Get-FileSystemAccessRule {
+    <#
+    .SYNOPSIS
+    Alternative to Get-Acl designed to be as lightweight and flexible as possible
+    TEMP NOTE: Get-DirectorySecurity combined with Get-FileSystemAccessRule replaces Get-FolderACE
+    .DESCRIPTION
+    Returns an object for each access control entry instead of a single object for the ACL
+    Excludes inherited permissions by default but allows them to be included with the -IncludeInherited switch parameter
+    .INPUTS
+    None. Pipeline input is not accepted.
+    .OUTPUTS
+    [PSCustomObject]
+    .NOTES
+    Currently only supports Directories but could easily be copied to support files, or Registry or AD providers
+    #>
+
+    param(
+
+        # Discretionary Access List whose FileSystemAccessRules to return
+        [System.Security.AccessControl.DirectorySecurity]$DirectorySecurity,
+
+        # Include inherited Access Control Entries in the results
+        [Switch]$IncludeInherited,
+
+        # Include non-inherited Access Control Entries in the results
+        [bool]$IncludeExplicitRules = $true,
+
+        # Type of IdentityReference to return in each ACE
+        [System.Type]$AccountType = [System.Security.Principal.SecurityIdentifier]
+
+    )
+
+    $AccessRules = $DirectorySecurity.GetAccessRules($IncludeExplicitRules, $IncludeInherited, $AccountType)
+    if ($AccessRules.Count -lt 1) {
+        Write-LogMsg @LogParams -Text "# Found no matching access rules for '$LiteralPath'"
+        return
+    }
+
+    $ACEPropertyNames = (Get-Member -InputObject $AccessRules[0] -MemberType Property, CodeProperty, ScriptProperty, NoteProperty).Name
+    ForEach ($ThisAccessRule in $AccessRules) {
+        $ACEProperties = @{
+            SourceAccessList = $SourceAccessList
+            Source           = 'Discretionary Access List'
+        }
+        ForEach ($ThisProperty in $ACEPropertyNames) {
+            $ACEProperties[$ThisProperty] = $ThisAccessRule.$ThisProperty
+        }
+        [PSCustomObject]$ACEProperties
+    }
+
+    <#
+    The creator of a folder is the Owner
+    Unless S-1-3-4 (Owner Rights) is in the DACL, the Owner is implicitly granted two standard access rights defined in WinNT.h of the Win32 API:
+      READ_CONTROL: The right to read the information in the object's security descriptor, not including the information in the system access control list (SACL).
+      WRITE_DAC: The right to modify the discretionary access control list (DACL) in the object's security descriptor.
+    Output an object for the Owner as well to represent that they have Full Control
+    #>
+    [PSCustomObject]@{
+        SourceAccessList  = $SourceAccessList
+        Source            = 'Ownership'
+        IsInherited       = $false
+        IdentityReference = $DirectorySecurity.Owner -replace '^O:', ''
+        FileSystemRights  = [System.Security.AccessControl.FileSystemRights]::FullControl
+        InheritanceFlags  = [System.Security.AccessControl.InheritanceFlags]::ContainerInherit -bor [System.Security.AccessControl.InheritanceFlags]::ObjectInherit
+        PropagationFlags  = [System.Security.AccessControl.PropagationFlags]::None
+        AccessControlType = [System.Security.AccessControl.AccessControlType]::Allow
+    }
+
+}