+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.67] - 2024-01-15 - added feature to included representative ACE for Owner in Get-FolderACE
+
+## [2.0.66] - 2024-01-15 - added feature to included representative ACE for Owner in Get-FolderACE
+
 ## [2.0.65] - 2024-01-15 - Resolve-Folder lns 35-37 workaround for https://github.com/IMJLA/PsNtfs/issues/1; also some DFS comments moved to PsDfs module
 
 ## [2.0.64] - 2024-01-13 - troubleshooting bug in get-folderace ln 51-55